extracted static methods and moved them to new class ImageDirectoryLocator
